requirements.txt
absl-py==2.2.2
asgiref==3.8.1
astunparse==1.6.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
django==5.2
django-allauth==65.7.0
flatbuffers==25.2.10
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
keras==3.9.2
libclang==18.1.1
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.5.1
namex==0.0.9
numpy==2.1.3
opt-einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
protobuf==5.29.4
pycparser==2.22
pygments==2.19.1
pyjwt==2.10.1
requests==2.32.3
rich==14.0.0
setuptools==80.2.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
typing-extensions==4.13.2
urllib3==2.4.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
